add helper for send notification
